get rid of EditComponent
